ref: Add versioning in bitbake file name.
