+++ common/recipes-lib/regli-retimer-v0.3.0/files/kb900x_utils.c
@@ -19,6 +19,16 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+void wait_ms(int milliseconds)
+{
+    clock_t start_time = clock();
+    // Convert milliseconds to clock ticks
+    const int nb_ms_in_sec = 1000;
+    clock_t wait_time = (milliseconds * CLOCKS_PER_SEC) / nb_ms_in_sec;
+    while (clock() < start_time + wait_time)
+        ; // Busy wait
+}
+
 uint8_t cal_crc8(uint8_t *data, size_t len)
 {
     uint8_t crc = 0x00;
