+++ common/recipes-lib/regli-retimer-v0.3.0/libregli-retimer_0.3.0.bb
@@ -3,10 +3,6 @@ SUMMARY = "kb900x Retimer Library"
 HOMEPAGE = "https://www.kandou.ai/"
 LICENSE = "Apache-2.0"
 LIC_FILES_CHKSUM = "file://${COREBASE}/meta/files/common-licenses/Apache-2.0;md5=89aea4e17d99a7cacdbeed46a0096b10"
-PV = "0.3.0"
-
-# Set default value (can be overridden in a .bbappend or local.conf)
-KB900X_BIC_COMMUNICATION ??= "True"
 
 S = "${WORKDIR}/sources"
 UNPACKDIR="${S}"
@@ -36,22 +32,8 @@ SRC_URI = "\
 
 inherit meson pkgconfig
 
-DEPENDS += ""
+DEPENDS += "libbic"
 
-RDEPENDS:${PN} += ""
+RDEPENDS:${PN} += "libbic"
 
-python __anonymous() {
-    bic_comm = d.getVar('KB900X_BIC_COMMUNICATION')
-    if bic_comm == "True":
-        deps = d.getVar('DEPENDS') or ""
-        rdeps = d.getVar('RDEPENDS') or ""
-        meson_opts = d.getVar("EXTRA_OEMESON") or ""
-
-        deps += " libbic"
-        rdeps += " libbic"
-        meson_opts += " -Dbic_communication=true"
-
-        d.setVar('DEPENDS', deps)
-        d.setVar('RDEPENDS', rdeps)
-        d.setVar("EXTRA_OEMESON", meson_opts)
-}
+EXTRA_OEMESON += "-Dbic_communication=true"
